// online-shop/src/components/interfaces.tsx
export interface ProductProps{
    id: number;
    category: string;
    name: string;
    description: string;
    price: number;
    picture: string;
  }
  
  export interface ProductState{
    id: number;
    category: string;
    name: string;
    description: string;
    price: number;
    picture: string;
  }